split compile mod from pattern mod in its own file
